iyb: for certain

+++ src/redblack.h
@@ -28,8 +28,8 @@
 #define RED 0
 #define BLACK 1
 //64bit architecture
-#define UNMAP_UNIT_POWER (size_t)12
-#define UNMAP_UNIT (size_t)(1<<UNMAP_UNIT_POWER)
+#define UNMAP_UNIT_POWER ((size_t)12)
+#define UNMAP_UNIT (size_t)((size_t)1<<UNMAP_UNIT_POWER)
 
 
 #define COLOR_BIT 1<<63
